set auto play amount

+++ features/content/play_widget/src/main/java/com/tokopedia/play/widget/data/PlayWidget.kt
@@ -69,7 +69,7 @@ data class PlayWidgetMeta(
     @SerializedName("gradient") val gradient: List<String> = emptyList(),
     @SerializedName("isAutoRefresh") val autoRefresh: Boolean = false,
     @SerializedName("autoRefreshTimer") val autoRefreshTimer: Long = 0,
-    @SerializedName("autoplayAmount") val autoplayAmount: Long = 0,
+    @SerializedName("autoplayAmount") val autoplayAmount: Int = 0,
     @SerializedName("autoplay") val autoplay: Boolean = false,
     @SerializedName("maxAutoplayCell") val maxAutoplayCell: Long = 0,
     @SerializedName("template") val template: String = ""
